requirements.txt
Django>=4.0
djangorestframework
psycopg2-binary
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-decouple